fic Speed and cont-item

+++ Kint-Sources-master/OpenNos.Domain/InventoryType.cs
@@ -14,7 +14,7 @@
 
 namespace OpenNos.Domain
 {
-    public enum InventoryType : byte
+    public enum InventoryType : short
     {
         Equipment = 0,
         Main = 1,
